+++ xxw-platform-applications/xxw-platform-datasource/xxw-platform-dynamic/src/main/java/com/xxw/platform/module/dynamic/entity/XxwWaybill.java
@@ -30,9 +30,9 @@ public class XxwWaybill implements Serializable {
 
     private Integer orderStatus;
 
-    private Byte shippingStatus;
+    private Integer shippingStatus;
 
-    private Byte payStatus;
+    private Integer payStatus;
 
     private String consignee;
 
@@ -50,7 +50,7 @@ public class XxwWaybill implements Serializable {
 
     private String postscript;
 
-    private Byte shippingId;
+    private Integer shippingId;
 
     private String shippingName;
 
@@ -139,19 +139,19 @@ public class XxwWaybill implements Serializable {
         this.orderStatus = orderStatus;
     }
 
-    public Byte getShippingStatus() {
+    public Integer getShippingStatus() {
         return shippingStatus;
     }
 
-    public void setShippingStatus(Byte shippingStatus) {
+    public void setShippingStatus(Integer shippingStatus) {
         this.shippingStatus = shippingStatus;
     }
 
-    public Byte getPayStatus() {
+    public Integer getPayStatus() {
         return payStatus;
     }
 
-    public void setPayStatus(Byte payStatus) {
+    public void setPayStatus(Integer payStatus) {
         this.payStatus = payStatus;
     }
 
@@ -219,11 +219,11 @@ public class XxwWaybill implements Serializable {
         this.postscript = postscript;
     }
 
-    public Byte getShippingId() {
+    public Integer getShippingId() {
         return shippingId;
     }
 
-    public void setShippingId(Byte shippingId) {
+    public void setShippingId(Integer shippingId) {
         this.shippingId = shippingId;
     }
